custom callback in a separate module
